separated pawn.cpp into a header and the corresponding file.
